athenaHLT: Add option to loop over input files
Add the `--loop-files` command line option. When set athenaHLT will loop
over the input files if the maximum number of events is larger than the
number of events in the input files(s). If no event count is given, it
will loop forever.

Also add a new options group for file I/O to make the help message look
a bit cleaner and fix arguments types for number of events.