Add Tool to implement electron ID based on a DNN
Adds AsgElectronSelectorTool (adapted from AsgElectronLikelihoodTool)
which retrieves variables and does some basic checks whether the object
is an electron etc. The TElectronSelectorTool applies all cuts placed
on variables including the DNN score. The DNN score is calculated in
the ElectronDNNTool using lwtnn. The ElectronSelectorTool is the parent
class of the ElectronDNNTool and enables to include other MVA methods
at a later point such as a CNN.