make data handles "functional" in 21.2
This should introduce a sort of limited functionality of the data
handles in 21.2, essentially I tried to make sure that everything that
worked in the unit test for the stand-alone handles also worked in
AthAnalysis.  This involved using the stand-alone decoration handles
in Athena as well, rather than trying to backport the full decoration
handles from master, which would likely run into issues with the
rather ancient version of gaudi in 21.2.  It also added the inline
constructors to the existing data handles for 21.2, to be consistent
with the standalone data handles.

This is not meant to work with AthenaMT, it is just meant to make it
easier to share code between 21.2 and master.  I'm also not trying to
guarantee that anything works that is not tested by the unit tests for
the dual-use handles.  I did add special warnings in several files to
make sure that they will not accidentally end up in master.