tauRecTools: support RNN ID for 2p taus
Hello,

This MR adds support for a dedicated 2p RNN ID network, as we are developing one for R22 (ATLTAU-1766, ATLTAU-1729).
So far we had networks for 1p, 3p (applied to >=2 prong), and 0p (trigger only).
The 0p and 2p networks are made optional, but the 1p and 3p are now mandatory (ERROR + FAILURE if not properly initialised).
I have verified that if no 2p network is given to the TauJetRNNEvaluator tool, the reconstruction output is unchanged.
Also, if we give the same network for 2p and 3p, the reconstruction output is unchanged, as expected.

To come: add support for efficiency flattening and working point calculation for 2p taus.

Cheers,
Bertrand