xAODTau: fix bug in TauDefs enum
Hello,

This MR is fixing a bug in a TauDefs enum.
The RNN electron veto working points were recently introduced in the IsTauFlag enum, at positions >=32.
However, we encode IsTauFlag on a bitset of 32 bits:
https://gitlab.cern.ch/atlas/athena/-/blob/master/Event/xAOD/xAODTau/Root/TauJet_v3.cxx#L274-287
So we cannot use anything higher than 31.
Consequently, I have moved the EleRNN WPs lower in the enum.

Currently, if we set EleRNNLoose=1, we read EleRNNLoose=0.
With the bug fix, the WPs work as intended.

The only change in the reconstruction output is the value of "isTauFlags" that encodes the tau WPs.
The RNN eVeto is new in R22, so there is no concern with backward compatibility.

Cheers,
Bertrand