tauRec: adjust q221 test setup
Hello,

This MR is reducing the number of events to process in our q221 test, because there is a huge bottleneck at the diff-root level.
Currently we reconstruct 500 events but diff-root 200 events, and this step alone takes about 13h...
We hope that reducing the number of reconstructed events to 200 will improve the situation, as diff-root probably needs to order the 500 events regardless of the number of events to compare.
If this is not good enough, we'll reduce the overall number of reconstructed events.

We also take this opportunity to drop the EMPFlow-seeded ART test, as it is not a viable configuration for the R22 reprocessing.

Cheers,
Bertrand