Remove TrigmuIso
Removing TrigmuIso as part of the trigger cleanup in ATR-22837. This code is partially obsolete Run 2 code, and partially code that has been migrated for Run 3. However, there is no plan to actually run this in Run 3 (it didn't run in Run 2 either), so it can all be removed. Have removed the remaining pieces of configuration from the Run 3 trigger, and the corresponding hypo alg/tool.