AthenaKernel: Fix potential race with extending mixed containers.
Updating the run/lbn fields of the ranges of the timestamp maps
for mixed containers is done in a separate step to the extension
itself.  So there is a window in which a lookup can return a range
with the old, unextended run/lbn.  Work around by retrying the lookup
if the range we get does not actually contain the sought key.

May address ATR-22831.