tauRec: final word on tau thinning
Hello,

This MR implements the final strategy for tau thinning in R22 :-)

So far, we were thinning TauPi0Clusters cells and TauShotClusters cells using CaloThinCellsByClusterAlg.
But in doing so, the cells were kept for all clusters in the input containers.
We could not keep only the cells for the subset of clusters.
However, in TauThinningAlg, we thin the TauPi0Clusters and TauShotClusters containers to remove clusters associated with low-pt taus.
Now, all tau-related thinning operations are done by TauThinningAlg, so we have full control over which clusters we are interested in, and accordingly, which cells we want to keep.
We also had a TauCellThinningAlg that was doing further cluster and cell thinning operations on CaloCalTopoClusters.
This has been moved to TauThinningAlg.

So we are left with a single tau thinning algorithm.

Thinning results have been validated (https://indico.cern.ch/event/1010330/contributions/4241055/attachments/2193441/3707715/Martin_TauRecoID_19Feb21.pdf), so TauThinningAlg is now running by default in ESDtoAOD.

Closes ATLTAU-1770.

Cheers,
Bertrand