tauRecTools: skip LRTs in tau track classification and downstream tools (ATLTAU-1772)
Hello,

In R22, we will have a dedicated reconstruction at xAOD level targeting Long Lived Particles decaying to taus, which will use Large Radius Tracks within 0.4 of the tau axis, and we'll have dedicated LRT tunes for tau algorithms (tau ID, ...).
For the R22 reprocessing already, we plan to associated LRTs with taus, so that LRTs can be retrieved via tau->tracks(xAOD::TauJetParameters::LargeRadiusTrack), which will facilite a lot the navigation at xAOD level.
However, we don't want the LRTs to participate in the standard tau reconstruction, besides the tau-track association.
This MR implements a mechanism to skip Large Radius Tracks in the tau track classifier and all downstream tools, even if LRTs have been associated with taus in TauTrackFinder.

The modification is made in the tau track classifier. If the tool is instructed to ignore LRTs (via tauRecFlags), it will filter out the LRTs possibly associated with the tau, decorate these with default RNN track classification scores, and proceed to classify the non-LRTs. I also fixed a problem with an incorrectly set xAOD::TauJetParameters::unclassified tau track flag.

The RNN tau ID tool now explicitly ignores the tracks that are not classified, so it is not affected by the possible presence of LRTs associated with taus in TauTrackFinder.
I have checked that the RNN ID score is unchanged when we associate LRTs with taus but don't classify them.
More validation will follow.

In TauAlgorithmsHolder, deprecated tools were moved to the end of the file, and will be cleaned up at the next occasion.

Cheers,
Bertrand