Adding support for truth-label-based JMS uncs
Some of the new large-R JMS uncertainties depend on the truth label,
as they are only applicable for QCD-labelled jets (not W/Z/top/etc
jets).  This MR adds support for such truth-label-based uncertainties
and also updates the validation plotting scripts for a few new types of
validation plots (done as part of the large-R consolidated tagger SFs
but not committed at the time).  This latter part is completely
irrelevant for users, but is useful for other developers of the package
who may need to validate new SF uncertainties in the future.