JetAlgorithm: Don't add duplicate decoration tool.
Don't add a track decoration tool if it's already done by a separate algorithm.

Actually, decoration should probably always be scheduled as a separate
algorithm, which the jet configuration can request.  That may work better
though with the new job configuration, so won't try to implement that
at this point.

See ATLASRECTS-6055 / ATLASRECTS-6057.