Concurrent processing of events in hybrid MP+MT workers
Modified `AthenaMPTools/SharedHiveEvtQueueConsumer` so that it can process
events concurrently. In order to achieve that this tool is now using
the `AthenaMtesEventLoopMgr` as the event processor, instead of the
previously used `AthenaHiveEventLoopMgr`.

Also introduced a helper abstract interface `AthenaKernel/IHybridProcessorHelper`
to allow the `SharedHiveEvtQueueConsumer` tool to call certain methods of the
`AthenaMtesEventLoopMgr`.