Remove check for container presence in CSC clusterization alg
Removig the 'isPresent' check for the CSC clusters. This was a left over from Run 2 trigger use cases and isn't needed since a) offline reco only runs this alg once per event and b) the trigger runs this alg once per view (and this check means that we only ran the clusterization in the first view we processed, which is the wrong behaviour).