jet hypo - reduce the load of prefiltering and condition filtering
Event prefiltering takes place in TrigJetHypoToolHelperNoGrouper,
and Condition prefiltering occurs before testing the leaf nodes in
FastReducer.

Often, the filtering is a "do nothing" operation, but is implemented
using an AcceptAll condition that runs on every jet. This is an expensive way to
do nothing. ConditionFilter has been modified to:

- do less copying
- recognise when it has no filtering to do.

The second is accomplished constructing FilterCondition with an empty Condition vector.

ConditionFilter now returns iterators into a vector of jets, rather than returning
a vector. In the case that it ahs no Conditions, it returns const begin and end
iterators of the input collection. Only if it has Conditions, does it copy the input
collection. Even here the copying is reduced the copy is partiioned by succesive
application of the Condtions, and the begin and new end iterators are returned
rather than retuening a copy of the partiioned vector.

The consequence of this is that the lifetime of the filter must be >= the lifetime
of its clients.