xAODRootAccess: Skip InheritsFrom tests that can't succeed.
In THolder, we were testing whether a type derives from SG::AuxVectorBase
or SG::AuxElement using TClass::InheritsFrom().
Change to test the STLtype first --- an AuxElement will be
kNoSTL and a DataVector will be kSTLlist.

This saves some time, and possibly autoparsing, which in turn avoids
a  cling assertion failure seen while running the unit tests in the
debug build.