tauRecTools: replace charged PFOs with tau tracks, and fix "bug/feature" in MVA TES
Hello,

In R22, we are not writing out tau charged PFOs in xAODs, because tau charged PFOs are supposed to be identical to tau tracks.
In this MR, we "safely replace" charged PFOs with tau tracks in one tool, so that the tool can be rerun on xAODs.

However... charged PFOs are actually only created for taus that have 1-5 tracks:
https://gitlab.cern.ch/atlas/athena/-/blob/master/Reconstruction/tauRecTools/src/TauPi0ClusterScaler.cxx#L29-44

This must have been overlooked in the MVA tau energy scale code, where "charged_totalE" is supposed to be the total energy carried by tau tracks, even when taus have more than 5 tracks...
The "charged_totalE" variable was null for taus with more than 5 tracks (these taus not used in analyses fortunately), and the behaviour is now fixed.
This MR does not change the reconstruction output for taus with at most 5 tracks.

Cheers,
Bertrand