First version of the SqliteReadSvc
The new Athena Service has been introduced to allow reading detector description
parameters from plain SQLite tables, which are collocated in the same SQLite
database together with the persistent representation of the GeoModel description
tree.

In order to allow for smooth transition of clients from the RDBAccessSvc to
the new SqliteReadSvc, the latter implements the IRDBAccessSvc interface.

The functionality that is yet to be implemented: an analogue of the RDBQuery for
reading data from SQLite tables.