xAODCore: Work around cling problems.
In some cases, some dictionary load requests will fail if they are done
in the `wrong' order.  This is a cling bug (as of root 6.20.00), but
we can work around it here.  We add a TClassGenerator so that if,
for example, we're trying to load ElementLink<T>, we first load T.
And similarly for a few other Athena template classes.
This is observed to be sufficient to fix the order-dependent failures.
See ATR-22681.

Also instantiate the BaseInfo specializations for containers in dictionaries.
This will be needed to work around other cling bugs which are causing
assertion failures in the dbg build.