Fix issues with the cost monitoring plottig
The automatic rangig caused extended the range beyong the "normal" range of
the histgrams for some histograms. Then the, frankly unwanted, root behaviour
of then including the underflow bin cause issues, since the histogram underflow
limit is always set to -1666.57.

This adds an additional check that prevents the range ever going below the
actual histogram displayed limit for the time.

Also, because the histograms now very sensibly, always have logarithmic
binning we now divide by the bin width by default.

This causes an issue with the computation of the means, so not the means
are calculated correctly using the integrals of the histograms rather than
the naive root calculation.