Update muon trigger configuration of SCT summary tool
Configuring the SCT summary tool to only configure the SCT_ByteStreamErrorsTool if we are looking at data. This prevents the muon reco tools from stalling due to the missing SCT byte stream errors data in MC (ATR-22926).

This also allows us to use the default configuration of the muon creator tool for combined muons which avoids a bunch of warnings from using the incorrect track summary tool.