Removing superfluous contents lists from DerivationFrameworkCore
This MR drops some contents lists that are now located in other packages but which were never removed from their original location in this package. This has caused confusion as to which lists are actually used, so removing them should reduce the confusion. The removed files were not in use and hence no changes to other parts of the software are needed.