Fix DAOD slimming bug which can cause dynamic EventInfo variables to be lost
ATLASDPD-1570

The EventInfo in the AOD is expressed both as static and dynamic aux variables. The static part is handled in CompulsoryContent.py but this does not address the dynamic part. Left alone all variables in EventInfo are still included via EventInfoAux. , but if a user erroneously tries to add an extra EventInfo item via ExtraVariables, this is the only dynamic variable that will make it into the output file (EventInfoAux.X) which is obviously not intended. This MR addresses this issue by adding a new list to CompulsoryContent into which containers that must be kept but which are expressed as dynamic variables in the AOD be added. The slimming machinery then adds these container names to AllVariables, meaning that (a) all variables are written and (b) if a user accidentally adds one of these variables as ExtraVariables, no harm is done because the overlap removal handles it (as with other AllVariables content). At the same time the extra EventInfo variable in DAOD_PHYS that revealed this bug is removed, as it is superfluous since all variables will be kept.