Introduced a common build_externals.sh script for the "external-less" projects.
This is meant to be used by AthDataQuality and DetCommon for the time
being. Just like build_project_externals.sh, meant for the other projects,
it requires a few simple settings from the individual build_externals.sh
scripts, and can then do the code checkout in a consistent way for both
of them.

At the same time fixed some small issues in the existing scripts as
well to ensure that code checkout errors would be properly reported
by them.