Adding allowed error to MadGraphControl
This error came up in testing, as diagnosed in AGENE-1962

It is associated with ghost script, which really has no impact at all on
MadGraph running in production (it has to do with writing images for
convenience for local users). This ensures that the job continues
happily past the error.