Using a much smaller minimum weight value in Pythia8
We have an old check for a "very small" weight, where that very
small value is fairly arbitrary. This seems to now be actually affecting
some BSM samples that are being generated.

Changing the check to use float minimum, because I fear these weights
are cast into floats at some point in the analysis chain even if they
aren't right here in this class. Also printing a warning in case the
weight is between double and float mins, to at least tell the user
what's happening a bit more clearly.