Adding alternative link path for truth helpers
This was already committed in 21.2; adding it here as pointed out in
ATLASG-1502 (I didn't realize it was missing, if I'm honest).

Basically, the truth link may have a different decoration name when
using some derivations. This lets the helper fall back to the
alternative name if the retrieval of the initial decoration fails.