First implementation of the mechanism for building GeoModel description from SQLite
Two new methods in the `IGeoModelSvc` interface:

1. `getParamSvcName()` returns the name of the Service which implements the `IRDBAccessSvc`
interface and which is used in the current job. This method allows us to avoid hardwiring the
name of the `IRDBAccessSvc` retrieved by the clients.
2. `buildFromSQLite()` to be used by `GeoModel` tools to determine whether or not the geometry
is being constructed from the SQLite file, and act accordingly.

Changes in the `GeoModelSvc`:

1. Introduced new property `SQLiteDB` which allows us to pass the path to the SQLite database file.
If the property is set, then the geometry is built from that SQLite file. Otherwise the geometry
is built from the Geometry Database tags.
2. Added a first implementation of the mechanism of building GeoModel description from the SQLite
database. This mechanism is not yet complete, but it can already be used for testing purposes.

Also fixed `SqliteRecord::dump()` so that it properly handles NULL values in table records.