Add QuasiStableParticleRadius SimFlag.
Particles decayed by the generator beyond this radius (mm) are
considered to be quasi-stable particles and will be passed to
Geant4 if quasi-stable particle simulation is activated.

Default value 30.19 mm. This is the largest radius before the IBL,
where there is a region with no material for most phi values.