Cleaned up the build of the PanTauAlgs package.
The main motivation was to make sure that the public headers of the
package would be properly installed in "Athena mode", in order for
the dictionary to work correctly.

But the cleanup grew into something more. I replaced the CINT
dictionary of the package into a Reflex one, introduced the
PanTauAlgsLib library that would be built in all circumstances,
cleaned up the dependencies of the package, and made sure that
the dictionary would refer to types with their correct names.

This is primarily meant to fix the issue described in ATEAM-721.