tauRecTools: replace Decorator with Accessor in TauTrackFinder
Hello,

As discussed in ATLASRECTS-6106 and ATLASRECTS-6104, we make an unnecessary and inappropriate use of Decorator on non-const objects, and we should have used Accessor instead.
Now fixed in TauTrackFinder.
Note for later: this could be a bit more streamlined by using the TauTrack::setDetail functionality of the EDM, extending the TrackDetail enum in TauDefs.h.

This MR does not change the reconstruction output.

Cheers,
Bertrand