tau PhysVal monitoring: add track and cluster variables
Hello,

This MR is adding a number of track and cluster variables to the tau PhysVal monitoring.
These are the variables used in RNN track classification, RNN tau ID, and MVA tau energy scale.
The new histograms were put semi-randomly in the "NewCorePlots" code, but the whole monitoring structure will have to be revised at some point.
The binning was significantly improved for a number of variables, and some broken variables were fixed.

More to come.

Cheers,
Bertrand