eflowRec: Fix memory `leak'
eflowTrackClusterLink was maintaining a pool of objects, looked up by
cluster and track pointers.  But these were never cleared, so these
objects built up over the entire run of the reconstruction.

Changed to use a SlotSpecificObj that gets cleared after an event is finished.
Also a number of other minor cleanups.