Fix for ALFA_SimHitsTestTool (ATLASSIM-5034)
`ALFA_SensitiveDetector` can make hits with `plate==100` for `GVS` volumes.
Usually `0<plate<10`. The hit plotting code in `ALFA_SimHitsTestTool` assumes
a maximum value of 10 when it uses `plate` as the index to arrays, so when a
hit with `plate==100` is processed this ends up well outside the array and
corrupts the stack. The fix is to put the `GVS` hits into a separate histogram,
skipping the stage where they are used as the index to an array.