PanTau: set decay mode to Mode_NotSet when not evaluated
Hello,

Currently, when taus have 0 or >5 associated tracks, the tau substructure reconstruction ("PanTau" tool) is not run, and the only substructure variable that is set for the tau is "PanTau_isPanTauCandidate = 0".
None of the other PanTau variables are set, and therefore, they all have a default value of 0.
This is very misleading for the PanTau "DecayMode" variable, for which a value of 0 corresponds to "Mode_1p0n" in the TauDefs enum.
The consequence is that "1p0n" taus currently includes 0p and >5p candidates.

In this MR, when PanTau_isPanTauCandidate=0, we also set the DecayMode to Mode_NotSet, to avoid this confusion.
We can now drop the workaround we added in TauCombinedTES.

The only change to the reconstruction output is TauJetsAux.PanTau_DecayMode: 0 -> 6, for 0p and >5p.

Closes ATLTAU-1775.

Cheers,
Bertrand