LArGeoBarrel code refactoring
Introduced new function LArGeo::buildElStraightSections() which builds
GeoStraightAccSection objects and records them into DetStore. This function will
be called by the LArDetectorTool when the GeoModel description is built from
the SQLite database. The code of this function has been extracted from
LArGeoBarrel/BarrelConstruction. As a result of this extraction, for the time
being there is large piece of duplicated code, which needs to be resolved later.

A set of auxiliary functions, which are used by the BarrelConstruction class and
the buildElStraightSections function, has been collected in a separate place
BarrelAuxFunctions(.h,.cxx)