Restructure FtagRun3DerivationConfig slightly
The idea is to return a component accumulator for each jet collection, so that
we can merge all of them together. There are still some points that need to be
cleaned up a bit, e.g. the calib service is just scheduled once, not in the
accumulator. Also it crashes right now...