Fix for ATLASRECTS-6000
In the case that MuGirl finds segments that include a small-large overlap, the list of hits it assembles to be fitted from the segments may be out of order.
This can cause problems for the hole-finding code, which assumes that the hits are in order.
Thus, once it reaches a new chamber it assumes that it has all the hits on track for the previous chamber.
Any hits that have been separated from the rest of the hits in the chamber are then treated as holes.
The hole-finding code searches for them, finds them, and adds them to the track, which then acquires duplicate hits.
In principle, MuGirl should prevent this from happening by sorting the hits in the case that there are small-large overlaps.
The MuonCandidateTrackBuilderTool checks for this by looping over the hits and seeing if both a small and a large one are present.
Unfortunately, at the moment it breaks out of the loop after the first MDT hit it finds.
With the break statement removed, the code now correctly identifies small-large overlaps and sorts the hits appropriately, so the track no longer acquires duplicate hits.
Additionally, the sorting code is modified to correctly handle pseudomeasurements.
Note that this will also fix many instances of ATLASRECTS-5999, even though this warning has now been downgraded.
Duplicate hits will also give rise to this warning, as it leads to a situation in which there are more hits on the track in a given chamber than the MuonTrackSummaryHelperTool finds close to the track in the chamber.