tauRec: set calibration path for all tools
Hello,

This MR is instructing tauRec to set the calibration path for all tools (even for tools that don't use calibration files) according to the tauRecToolsCVMFSPath flag.
Currently, if not specified explicitly in TauAlgorithmsHolder, we are picking up some default path set in TauRecToolBase (the default has been changed to an empty string, to make sure we don't pick a wrong default value), and the path must be set for each tool individually, which is error prone.

The only exception that still remains is the RNN tau track classifier, where the path is set for the tool of a tool, rather than the tool directly. This is to be improved.

I also cleaned up deprecated tool configurations and flags that will no longer be used in R22.

This MR does not change the reconstruction output.

Cheers,
Bertrand