tauRecTools: set back default path in TauRecToolBase
Hello,

Apparently, !42357 removed something that client codes rely on: the default path for calibration tau files.
That is mostly be an issue for analysis tools, rather than a problem for reconstruction.
But for the time being, I will put back the default path, until this gets clarified and improved.

Adding the "urgent" label, because unit tests are currently failing due to this problem.

Sorry for the mess.
Bertrand