tauRec: sort final taus by pt
Hello,

This MR sorts the final tau candidates by decreasing pt (where pt = ptFinalCalib = MVA pt), which is a bit nicer for the user.
Until now, the ordering was given by the LCTopo seed jet order (we loop over seed jets and create a tau for every seed jet, and we proceed sequentially, so we keep the seed jet order).

Cheers,
Bertrand