fix AsgServiceConfig/AnaAlgorithmConfig python unit test failure
Those tests were broken by !42411, as noted in ATLASG-1589.  This
change should fix that.

The specific problem was that the C++ code now does its own checks
whether subtools are defined, and throws its own exceptions.  Those
differ from the exceptions python throws and the unit test expects.
There are several ways to fix this, but it seemed most reasonable to
first do the checks in python before trying to register subtools in
the C++ code.