Remove or demote excessive printout
Excessive WARNINGS from TileMuonReceiverDecision and TGCTriggerDBAlg have been demoted.

In the case of TileMuonReceiverDecision, it demotes the output after the first 20 WARNINGS
in the loop, such that the first 20 are WARNINGS, and the subsequent WARNINGS are printed
as DEBUG.

Fro the TCG output, since the information in the channel was not being printed, then all
the INFO lines were identical, which is a bit excessive - unless the objective is simply
to spot any which are *not* identical, in which case, it would be better to simply
do a comparison of them all, and then print out the number that are different.