Run both iterative and adaptive vertices in the FS tracking
Changes to run both iterative and adaptive vertex algorithms in the FS tracking from the jets,
including factorising out the TTVA such that we can run separate instances of the TTVA for each
algorithm.

This also changes the default vertex algorithm to be the adaptive, so that the counts will change

- in principle the purpose of this change is to allow the jets to still use the iterative algorithm,
so some further changes will be necessary in the code to switch to use this different vertex.