TauDQA: protect tau PhysVal monitoring against missing vertex container
Hello,

This MR adds a protection in the tau PhysVal monitoring against missing PrimaryVertices container.
This happens in our standalone tauRec ART tests, where PrimaryVertices is not in the output AOD, and we run the PhysVal monitoring on this AOD, which crashes because the element link is invalid.
The problem appeared recently, when we added tau vertex histograms to the monitoring, in !41977 .
It's not a problem for PhysVal checks, where regular AODs are used.

Cheers,
Bertrand