Bugfix for CombinedMuonTrackBuilder
Due to a typo, the track was not being correctly checked after the error optimization step.
Since this is the last step of the fitting, this means that problematic tracks could potentially make it through.
As the entire track is not rejected, simply the result of the final optimization step, the impact of this change should be limited.
However, it should remove the FPEs due to negative covariance matrix entries reported in ATLASRECTS-6198.