fix AnalysisBase grid submission problems introduced with !42411
The changes introduced in !42411 required additional dictionaries to
serialize any of the component configuration objects, which were not
defined in that merge request.  This is (apparently) not caught in any
unit tests, but it leads to any sub-tools to be stripped off
algorithms during batch and grid submission.

This MR should fix that by defining one dictionary and changing the
type of one member variable to avoid the need for the second.  At some
point a new test should be defined to catch these problems in CI, or
at least in the nightly tests, but for now having a quick fix seemed
more important.