Remove muon dR test trigger
Removing the muon Dr trigger from the menu. This was only meant as a test example for applying combined cuts on multiple muons, and never planned to be used. We have real chains that cover the use case this was testing (cutting on the invariant mass of the muon pairs).