tauRec: use "acmd.py diff-root" instead of "art.py compare" until "--branches-of-interest" is supported
Hello,

As explained in ATLINFR-4057, it seems that the "--branches-of-interest" option is not yet supported when using "art.py compare", while it is functional when using the underlying command "acmd.py diff-root". This is causing all tauRec ART diff-root commands to fail, so I'm temporarily changing the command, until this option is supported.

Adding the "urgent label" as having a functional "diff-root" in our tests is really helpful.

Cheers,
Bertrand