First and incomplete version of LArDetectorFactoryLite
This commit is an intermediate step towards being able to construct a complete
LAr GeoModel description in Athena from an SQLite database.

* Introduced a first version of the LArDetectorFactoryLite class, which is
invoked by LArDetectorTool when the geometry description is built from SQLite.
For the time being the new factory only builds barrel electrode straight section
objects (for Fast Sim), and records StoredAlignX and StoredPhysVol objects into
detector store.
* LArDetectorTool has been updated so that it uses the LArDetectorFactoryLite when
the geometry is built from SQLite, and the old LArDetectorFactory when the geometry
is built from the GeometryDB. Also, the Tool does not attempt to construct LArHVManager
when the geometry is built from SQLite. This is a temporary measure until the
issue with initializing ID dictionaries from SQLite input gets sorted out.
* Added two new methods to the IGeoDbTagSvc interface to allow subsystem detector
tools to get a pointer to the ReadGeoModel object (for retrieving subsystem-specific
maps of Full Physical Volumes and Alignable Transforms), and to allow all clients
of the IRDBAccessSvc interface to read the parameters through the SqliteReadSvc when
the geometry is built from SQLite.