tauRecTools: fix definition of array of array
Hello,

This MR fixes the definition of arrays of arrays in a tau pt calculation.
The source code uses m_array[decayModeIndex][etaIndex], while in the header, we declare std::array<std::array<XXX, DecayModeBinning>, EtaBinning> m_array with swapped indices.
By chance, DecayModeBinning = EtaBinning = 5, that's why we don't see any problem...

This MR is a bug fix, but in practice, it does not change the reconstruction output.

Cheers,
Bertrand