add a addPrivateToolInArray function to DualUseConfig.py
Unfortunately I don't have a really good way to test if it works, but
the underlying C++ is tested, and this is hopefully a close enough
duplication of what's happening for non-array tools that this is good
enough.

Originally I tried to do something more involved, but that became sort
of tedious given the logic of the python class.  Maybe at some point I
will come back and restructure the python class somewhat.  Hopefully
this is still workable as is.