Add extra sanity check for combined tracks
In the rare case that the track fit fails after the Eloss update and MS-ID alignment steps, the code reverts to the original track.
However, in such a case the basic track quality checks are not applied to that original track, which is instead passed directly to the hole recovery step.
This fix simply rejects the track prior to hole recovery if it fails these checks: as such tracks are very poor quality anyway, this should not be an issue.
This should also allow us to avoid some FPEs coming from covariance matrices with negative diagonal elements.