Support for a new generic histogram scaling in jet calibration
This adds a new GenericHistScaleCorrection class, which is designed
originally with the trigger-level dijet analysis in mind (where there is
the need for an additional online --> offline calibration factor), but
which can additionally be used also in other contexts as a generic way to
insert another calibration step into the jet calibration chain.

For the moment, as there is only one use case, the code is aligned
around the idea of a 1D histogram where the x-axis is pT, being applied
to a jet where the pT is scaled to a new value.  However, additional
configuration options could be added to easily adjust this to also
consider other histogram dimensionality, axis definitions, or methods
for scaling the jet four-vector.  These are not done at this time
because it's not clear how much this is needed beyond the TLA, rather the
code is designed to be more modular to support such extensions without too
much additional effort.