tauRec: tidy up calibration flags
Hello,

This MR is defining a few more tauRec flags to avoid hardcoding calibration files in TauAlgorithmsHolder. It makes it easier to change the files via a preExec.
We also move the default calibration path to "R22_preprod", anticipating the upcoming MRs that will deploy the R22 preliminary tunes to be used for Round2 MC pre-production.

Cheers,
Bertrand