Use same RoIs for L2 and EF combined muons
As pointed out in ATR-23252, it would be better to use the same RoIs for the fast and precision tracking and combined muon building steps. To do this, have added a link to the RoI in the L2 combined muon hypo alg, and switched the EF combined muon input maker to use the NamedRoITool instead of just fetching the RoI from the previous step (the EF SA muon rois from the previous step are similar but not quite identical to the combined muon RoIs).