tauRecTools: deploy R22 preliminary tune for RNN track classifier
Hello,

This MR deploys the R22 preliminary tune for RNN track classification derived on Round1 MC pre-prod samples (ATLTAU-1738).
The RNN uses additional variables, some related to the primary vertex, which necessitates to read in the PrimaryVertices container, and pass it on to the downstream tool that actually evaluates the RNN inputs.
Some minor code cleanup and reorganisation was also made.

Adding the urgent flag as we need this MR for Round2 MC pre-production to be launched very soon.
We might readjust the input variables for the Run2 reprocessing, based on final studies to be made on Round2 samples.

Cheers,
Bertrand