tauRecTools: deploy preliminary R22 tune for calo TES and MVA TES
Hello,

This MR is deploying the preliminary R22 tunes for calo TES (ATLTAU-1744) and MVA TES (ATLTAU-1739) derived on Round1 MC pre-prod samples.
For calo TES, it's only a change in calibration file.
For MVA TES, to support the new calibration file, we need to add the 2 new input variables for the BRT: the median event energy density rho, and the ratio between EM and LC scale pt.
That results in 2 new decorations added onto the tau, "rho" and "ptIntermediateAxisEM". The former is kept in the xAOD output for the time being, to ease the validation, but will be dropped for the reprocessing.

Cheers,
Bertrand