Fix for crash due to missing roi link in muon and bphysics triggers
Fixing ATR-23338. Adding an roi link in the BPhysics hypo at the L2 combined stage similar to what is done for the muComb hypo so that we have the roi link available at the EF combined muon step even if there are only Bphysics chains active for a particular view. Also adding an exception for the roi links to avoid warnings about the link being found by multiple paths (as it is expected for this case).