Remove old NSW hack from MuonSeededSegmentFinder
Prior to the existence of a RegSelTool for the NSW, a hack was added here to allow PRDs to be retrieved.
However, this is no longer necessary now that the NSW RegSelTools are available.
Additionally, by incorrectly indicating to the code to search for holes in all available chambers, it can result in duplicate hits being placed on tracks.
This in turn can lead to the reconstruction of extra muons, as the duplicate hits may cause the overlap removal to treat the combined track and the original extrapolated track as different.
This should hopefully fix the problem observed in ATLASRECTS-5610.