Put in a noddy han config, which may, or may not, work
This is the habn confg made using one of the test runs provided by the DQ group.
it includes fragments from one of the other han configs.

By necessity, the menu was changing, so using this approach is really very flawed
since it means that we will have histograms included which will not be enabled for
all the runs that are being used.

This is extremely poor practice, as it means the dhift people will hav e no way to
tell when a chain is not present because it was not in the menu, or because of some
genuine problem with the trigger or the monitoring.